feat(manga): restructure MangaControllerV1 and introduce MangaQuery and MangaFilter for enhanced query capabilities

+++ src/main/java/com/manga/server/features/manga/dtos/MangaDTO.java
@@ -1,4 +1,4 @@
 package com.manga.server.features.manga.dtos;
 
-public record MangaDTO(String id, String name, String thumbnail ,Double lastChapter, String description) {
+public record MangaDTO(String id, String name, String thumbnail, Integer lastChapter, String description) {
 };
